Included DHT22 libraries to read temperature. Fixed humidity (was int, but DHT22 returns float).

+++ NanodeRF_Cosm_Jarduino/NanodeRF_Cosm_Jarduino.ino
@@ -56,7 +56,7 @@
 typedef struct 
 { int jardinera; 
   float temperature;
-  int humidity; 
+  float humidity; 
   int soilMoisture; 
   int sunlight;
   int angle;
